+++ src/parse-dns.h
@@ -19,7 +19,7 @@
      *  @param bytes pointer to the packet
      *  @param packet_offset_size current offset in the packet
      */
-    void parse_dns(const u_char* bytes, size_t packet_offset_size);
+    void parse_dns(const u_char* bytes, int32_t packet_offset_size);
 
     /**
      *  @brief Parse a Question in the Questions section on the DNS frame of the packet
@@ -27,7 +27,7 @@
      *  @param bytes pointer to the packet
      *  @param packet_offset_size current offset in the packet
      */
-    size_t parse_dns_question(const u_char* bytes, size_t packet_offset_size);
+    size_t parse_dns_question(const u_char* bytes, int32_t packet_offset_size);
 
     /**
      *  @brief Parse an Answer in the Answers section in the DNS frame of the packet
@@ -35,7 +35,7 @@
      *  @param bytes pointer to the packet
      *  @param packet_offset_size current offset in the packet
      */
-    size_t parse_dns_answer(const u_char* bytes, size_t packet_offset_size);
+    int32_t parse_dns_answer(const u_char* bytes, int32_t packet_offset_size);
     
     /**
      *  @brief Parse the label/ptr name field in the DNS frame
@@ -47,7 +47,7 @@
      * 
      *  @return updated packet_offset_size
      */
-    size_t parse_dns_name_field(const u_char* bytes, size_t packet_offset_size, std::string &name, bool is_pointer_reference);
+    size_t parse_dns_name_field(const u_char* bytes, int32_t packet_offset_size, std::string &name, bool is_pointer_reference);
 
     /**
      *  @brief Parse one or more <character-string>s and save them into a string
@@ -58,7 +58,7 @@
      * 
      *  @return updated packet_offset_size
      */
-    size_t parse_dns_string(const u_char* bytes, size_t packet_offset_size, std::string &name);
+    size_t parse_dns_string(const u_char* bytes, int32_t packet_offset_size, std::string &name);
 
     /**
      *  @brief Parse the Record Data field in the DNS frame 
@@ -69,7 +69,7 @@
      *  
      *  @return parsed RDATA
      */
-    std::string parse_dns_answer_rdata(const u_char* bytes, size_t packet_offset_size, uint16_t rdata_length, uint16_t type);
+    std::string parse_dns_answer_rdata(const u_char* bytes, int32_t packet_offset_size, uint16_t rdata_length, uint16_t type);
     
     /**
      *  @brief Converts integer DNS type to a readable string format
@@ -80,13 +80,13 @@
      */
     std::string dns_type_to_string(uint16_t type);
 
-    std::string parse_dns_answer_rdata_soa(const u_char* bytes, size_t packet_offset_size);
-    std::string parse_dns_answer_rdata_mx(const u_char* bytes, size_t packet_offset_size);
+    std::string parse_dns_answer_rdata_soa(const u_char* bytes, int32_t packet_offset_size);
+    std::string parse_dns_answer_rdata_mx(const u_char* bytes, int32_t packet_offset_size);
 
     // DNSSEC 
-    std::string parse_dns_answer_rdata_rrsig(const u_char* bytes, size_t packet_offset_size, uint16_t rdata_length);
-    std::string parse_dns_answer_rdata_dnskey(const u_char* bytes, size_t packet_offset_size, uint16_t rdata_length);
-    std::string parse_dns_answer_rdata_nsec(const u_char* bytes, size_t packet_offset_size, uint16_t rdata_length);
-    std::string parse_dns_answer_rdata_ds(const u_char* bytes, size_t packet_offset_size, uint16_t rdata_length);
+    std::string parse_dns_answer_rdata_rrsig(const u_char* bytes, int32_t packet_offset_size, uint16_t rdata_length);
+    std::string parse_dns_answer_rdata_dnskey(const u_char* bytes, int32_t packet_offset_size, uint16_t rdata_length);
+    std::string parse_dns_answer_rdata_nsec(const u_char* bytes, int32_t packet_offset_size, uint16_t rdata_length);
+    std::string parse_dns_answer_rdata_ds(const u_char* bytes, int32_t packet_offset_size, uint16_t rdata_length);
     
 #endif // PARSE_DNS_H_
